Inlining DynamicParameterListBuilder
